Fixed directional convention of energy flows in thermal_simulation.py
The calculation of the mass flow out of nodes was based on calculating the mass flow into nodes and flipping the sign. Now the mass flow out of nodes is calculated directly. Given conservation of mass, this should not change anything. However, there are small numerical imprecisions. This change make the calculation more accurate.